% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ads.R
\name{get_ads}
\alias{get_ads}
\title{Get Ads Report.}
\usage{
get_ads(start_date = NULL, end_date = NULL, group_by = NULL,
  granularity = NULL, products = NULL, countries = NULL,
  networks = NULL, format = NULL, output = NULL, parse_as = NULL)
}
\description{
Get Ads Report.
}
