% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sales.R
\name{get_regions_sales}
\alias{get_regions_sales}
\title{Get Region Sales Report}
\usage{
get_regions_sales(start_date = NULL, end_date = NULL, group_by = NULL,
  granularity = NULL, products = NULL, countries = NULL, dataset = NULL,
  include_inapps = NULL, format = NULL, output = NULL, parse_as = NULL)
}
\description{
Get Region Sales Report
}
