% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_revenue.R
\name{get_revenue}
\alias{get_revenue}
\title{Get Revenue Report.}
\usage{
get_revenue(start_date = NULL, end_date = NULL, group_by = NULL,
  products = NULL, countries = NULL, dataset = NULL,
  include_inapps = NULL, format = NULL, output = NULL, parse_as = NULL)
}
\description{
Get Revenue Report.
}
