% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sales.R
\name{get_sales}
\alias{get_sales}
\title{Get Sales Report.}
\usage{
get_sales(start_date = NULL, end_date = NULL, group_by = NULL,
  granularity = NULL, products = NULL, countries = NULL, dataset = NULL,
  include_inapps = NULL, format = NULL, output = NULL, parse_as = NULL)
}
\description{
Get Sales Report.
}
